requirements.txt
pyserial==3.5
pydantic~=2.6.1